requirement.txt
Click==7.0
colorama==0.4.0
Django==2.1.2
django-cors-headers==2.4.0
django-tastypie==0.14.2
django-thrift==0.1.3
django-vue==0.0.1
djangorestframework==3.9.0
djangovue==18.11.18
gevent==1.2.2
greenlet==0.4.15
gunicorn==19.9.0
gunicorn-thrift==0.2.22
happybase==1.1.0
ply==3.11
python-dateutil==2.7.3
python-mimeparse==1.6.0
python-vuejs==1.0.9
pytz==2018.5
setproctitle==1.1.10
six==1.11.0
thrift==0.11.0
thriftpy==0.3.9
vue==0.0.1
